requirements.txt
pytest
psutil
memory_profiler
pydantic
python-dotenv
requests
structlog
tqdm